sns_functions taken out of views
